+++ src/features/dashboard/settings/orderSource/OrderSourcePage.jsx
@@ -1,6 +1,7 @@
 import React, { useEffect, useState } from "react";
 import DataTable, { defaultThemes } from "react-data-table-component";
 import { useDispatch, useSelector } from "react-redux";
+import { toast } from "react-toastify";
 import { openModal } from "../../../../app/modal/modalReducer";
 import { deleteOrderSource, loadOrderSource } from "./orderSourceActions";
 
@@ -179,7 +180,15 @@ export default function OrderSourcePage() {
             data-name="delete"
             id={orderSource.id}
             onClick={() => {
-              dispatch(deleteOrderSource(orderSource.id));
+              if (
+                prompt(`Zəhmət olmasa silmək üçün şifrəni daxil edin`) == 9519
+              ) {
+                dispatch(deleteOrderSource(orderSource.id));
+              } else {
+                toast.info(
+                  "Silmək cəhtiniz uğursuzdur, silmək üçün düzgün şifrə daxil edin."
+                );
+              }
             }}
             onMouseEnter={(e) => {
               sethover(true);
