+++ src/features/dashboard/settings/reference/ReferencePage.jsx
@@ -1,6 +1,7 @@
 import React, { useEffect, useState } from "react";
 import DataTable, { defaultThemes } from "react-data-table-component";
 import { useDispatch, useSelector } from "react-redux";
+import { toast } from "react-toastify";
 import { openModal } from "../../../../app/modal/modalReducer";
 import { deleteReference, loadReference } from "./referenceActions";
 
@@ -177,7 +178,15 @@ export default function ReferencePage() {
             data-name="delete"
             id={reference.id}
             onClick={() => {
-              dispatch(deleteReference(reference.id));
+              if (
+                prompt(`Zəhmət olmasa silmək üçün şifrəni daxil edin`) == 9519
+              ) {
+                dispatch(deleteReference(reference.id));
+              } else {
+                toast.info(
+                  "Silmək cəhtiniz uğursuzdur, silmək üçün düzgün şifrə daxil edin."
+                );
+              }
             }}
             onMouseEnter={(e) => {
               sethover(true);
